refactor(batch-exports): Move get_object_key tests to shared test_utils
- Move S3 key generation tests from s3/test_utils.py to shared test_utils.py
- Update README to include batch-exports compose file in docker command
